refactor: notifications route
