--- src/app/(normal)/notifications/layout.tsx
@@ -1,15 +1,0 @@
-import { t } from '@lingui/macro';
-import type React from 'react';
-
-import { createPageTitleSSR } from '@/helpers/createPageTitle.js';
-import { createSiteMetadata } from '@/helpers/createSiteMetadata.js';
-
-export async function generateMetadata() {
-    return createSiteMetadata({
-        title: createPageTitleSSR(t`Notifications`),
-    });
-}
-
-export default function DetailLayout({ children }: { children: React.ReactNode }) {
-    return <>{children}</>;
-}
